refactor(mcp): decompose lib.rs into focused submodules (KAN-765)
