+++ crates/kanban-mcp/src/helpers/mod.rs
@@ -1,0 +1,15 @@
+pub(crate) mod error_mapping;
+pub(crate) mod macros;
+pub(crate) mod parsers;
+pub(crate) mod resolvers;
+pub(crate) mod serialization;
+
+pub(crate) use error_mapping::{
+    core_err_to_mcp, kanban_err_to_mcp, mcp_enrich_add_error, mcp_enrich_remove_error,
+};
+pub(crate) use macros::{locked_read, locked_write, mutating_op, read_op};
+pub(crate) use parsers::{
+    parse_datetime, parse_priority, parse_sort_field, parse_sort_order, parse_status,
+};
+pub(crate) use resolvers::{card_board, resolve_summaries, McpResolve};
+pub(crate) use serialization::{to_call_tool_result, to_call_tool_result_json};
