+++ crates/kanban-mcp/src/tools/mod.rs
@@ -1,0 +1,5 @@
+pub(crate) mod board;
+pub(crate) mod card;
+pub(crate) mod column;
+pub(crate) mod sprint;
+pub(crate) mod transfer;
